+++ inlong-dashboard/src/configs/menus/conf.tsx
@@ -101,10 +101,6 @@ const conf: MenuItemType[] = [
         path: '/system',
         name: i18n.t('configs.menus.ModuleAudit'),
       },
-      {
-        path: '/systemId',
-        name: i18n.t('configs.menus.ModuleAuditID'),
-      },
     ],
   },
 ];
