+++ inlong-dashboard/src/i18n.ts
@@ -37,8 +37,7 @@ const resources = {
       'configs.menus.DataSynchronize': 'Synchronization',
       'configs.menus.TenantManagement': 'Tenant Management',
       'configs.menus.SystemOperation': 'Operation',
-      'configs.menus.ModuleAudit': 'Module audit (IP)',
-      'configs.menus.ModuleAuditID': 'Module audit (ID)',
+      'configs.menus.ModuleAudit': 'Module audit',
     },
   },
   cn: {
@@ -55,8 +54,7 @@ const resources = {
       'configs.menus.DataSynchronize': '数据同步',
       'configs.menus.TenantManagement': '租户管理',
       'configs.menus.SystemOperation': '系统运维',
-      'configs.menus.ModuleAudit': '模块审计 (IP)',
-      'configs.menus.ModuleAuditID': '模块审计 (ID)',
+      'configs.menus.ModuleAudit': '模块审计',
     },
   },
 };
