+++ inlong-dashboard/src/ui/pages/ModuleAudit/IdModule/config.tsx
@@ -20,29 +20,20 @@
 import dayjs from 'dayjs';
 import i18n from '@/i18n';
 
-export const toChartData = (source, sourceDataMap) => {
-  const xAxisData = Object.keys(sourceDataMap);
-  return {
-    legend: {
-      data: source.map(item => item.auditName),
-    },
-    tooltip: {
-      trigger: 'axis',
-    },
-    xAxis: {
-      type: 'category',
-      data: xAxisData,
-    },
-    yAxis: {
-      type: 'value',
-    },
-    series: source.map(item => ({
-      name: item.auditName,
-      type: 'line',
-      data: xAxisData.map(logTs => sourceDataMap[logTs]?.[item.auditId] || 0),
-    })),
-  };
-};
+export const timeStaticsDimList = [
+  {
+    label: i18n.t('pages.GroupDetail.Audit.Min'),
+    value: 'MINUTE',
+  },
+  {
+    label: i18n.t('pages.GroupDetail.Audit.Hour'),
+    value: 'HOUR',
+  },
+  {
+    label: i18n.t('pages.GroupDetail.Audit.Day'),
+    value: 'DAY',
+  },
+];
 
 export const toTableData = (source, sourceDataMap) => {
   return Object.keys(sourceDataMap)
@@ -53,11 +44,60 @@ export const toTableData = (source, sourceDataMap) => {
     }));
 };
 
-export const getFormContent = (initialValues, onSearch) => [
+export const getFormContent = initialValues => [
   {
-    type: 'inputsearch',
-    label: i18n.t('pages.ModuleAuditDashboard.config.Ip'),
-    name: 'ip',
+    type: 'select',
+    label: i18n.t('pages.ModuleAudit.config.InlongGroupId'),
+    name: 'inlongGroupId',
+    props: {
+      dropdownMatchSelectWidth: false,
+      options: {
+        requestAuto: true,
+        requestService: {
+          url: '/group/list',
+          method: 'POST',
+          data: {
+            pageNum: 1,
+            pageSize: 1000,
+            inlongGroupMode: 0,
+          },
+        },
+        requestParams: {
+          formatResult: result =>
+            result?.list.map(item => ({
+              label: item.inlongGroupId,
+              value: item.inlongGroupId,
+            })) || [],
+        },
+      },
+    },
+  },
+  {
+    type: 'select',
+    label: i18n.t('pages.ModuleAudit.config.InlongStreamId'),
+    name: 'inlongStreamId',
+    props: values => ({
+      dropdownMatchSelectWidth: false,
+      options: {
+        requestAuto: true,
+        requestService: {
+          url: '/stream/list',
+          method: 'POST',
+          data: {
+            pageNum: 1,
+            pageSize: 1000,
+            inlongGroupId: values.inlongGroupId,
+          },
+        },
+        requestParams: {
+          formatResult: result =>
+            result?.list.map(item => ({
+              label: item.inlongStreamId,
+              value: item.inlongStreamId,
+            })) || [],
+        },
+      },
+    }),
   },
   {
     type: 'datepicker',
@@ -67,7 +107,7 @@ export const getFormContent = (initialValues, onSearch) => [
     props: {
       allowClear: false,
       showTime: true,
-      format: 'YYYY-MM-DD HH:MM:ss',
+      format: 'YYYY-MM-DD HH:mm:ss',
     },
   },
   {
@@ -78,12 +118,12 @@ export const getFormContent = (initialValues, onSearch) => [
     props: {
       allowClear: false,
       showTime: true,
-      format: 'YYYY-MM-DD HH:MM:ss',
+      format: 'YYYY-MM-DD HH:mm:ss',
     },
   },
   {
     type: 'select',
-    label: i18n.t('pages.ModuleAuditDashboard.config.BenchmarkIndicator'),
+    label: i18n.t('pages.ModuleAudit.config.BenchmarkIndicator'),
     name: 'benchmark',
     props: {
       allowClear: true,
@@ -106,7 +146,7 @@ export const getFormContent = (initialValues, onSearch) => [
   },
   {
     type: 'select',
-    label: i18n.t('pages.ModuleAuditDashboard.config.ComparativeIndicators'),
+    label: i18n.t('pages.ModuleAudit.config.ComparativeIndicators'),
     name: 'compared',
     props: {
       allowClear: true,
@@ -137,16 +177,8 @@ export const getTableColumns = source => {
   }));
   return [
     {
-      title: i18n.t('pages.GroupDetail.Audit.Time'),
-      dataIndex: 'logTs',
-    },
-    {
-      title: i18n.t('pages.ModuleAuditDashboard.config.InlongGroupId'),
-      dataIndex: 'inlongGroupId',
-    },
-    {
-      title: i18n.t('pages.ModuleAuditDashboard.config.InlongStreamId'),
-      dataIndex: 'inlongStreamId',
+      title: i18n.t('pages.ModuleAudit.config.Ip'),
+      dataIndex: 'ip',
     },
   ].concat(data);
 };
