+++ inlong-dashboard/src/ui/pages/ModuleAudit/IpModule/config.tsx
@@ -20,20 +20,29 @@
 import dayjs from 'dayjs';
 import i18n from '@/i18n';
 
-export const timeStaticsDimList = [
-  {
-    label: i18n.t('pages.GroupDetail.Audit.Min'),
-    value: 'MINUTE',
-  },
-  {
-    label: i18n.t('pages.GroupDetail.Audit.Hour'),
-    value: 'HOUR',
-  },
-  {
-    label: i18n.t('pages.GroupDetail.Audit.Day'),
-    value: 'DAY',
-  },
-];
+export const toChartData = (source, sourceDataMap) => {
+  const xAxisData = Object.keys(sourceDataMap);
+  return {
+    legend: {
+      data: source.map(item => item.auditName),
+    },
+    tooltip: {
+      trigger: 'axis',
+    },
+    xAxis: {
+      type: 'category',
+      data: xAxisData,
+    },
+    yAxis: {
+      type: 'value',
+    },
+    series: source.map(item => ({
+      name: item.auditName,
+      type: 'line',
+      data: xAxisData.map(logTs => sourceDataMap[logTs]?.[item.auditId] || 0),
+    })),
+  };
+};
 
 export const toTableData = (source, sourceDataMap) => {
   return Object.keys(sourceDataMap)
@@ -44,60 +53,11 @@ export const toTableData = (source, sourceDataMap) => {
     }));
 };
 
-export const getFormContent = initialValues => [
+export const getFormContent = (initialValues, onSearch) => [
   {
-    type: 'select',
-    label: i18n.t('pages.ModuleAuditDashboard.config.InlongGroupId'),
-    name: 'inlongGroupId',
-    props: {
-      dropdownMatchSelectWidth: false,
-      options: {
-        requestAuto: true,
-        requestService: {
-          url: '/group/list',
-          method: 'POST',
-          data: {
-            pageNum: 1,
-            pageSize: 1000,
-            inlongGroupMode: 0,
-          },
-        },
-        requestParams: {
-          formatResult: result =>
-            result?.list.map(item => ({
-              label: item.inlongGroupId,
-              value: item.inlongGroupId,
-            })) || [],
-        },
-      },
-    },
-  },
-  {
-    type: 'select',
-    label: i18n.t('pages.ModuleAuditDashboard.config.InlongStreamId'),
-    name: 'inlongStreamId',
-    props: values => ({
-      dropdownMatchSelectWidth: false,
-      options: {
-        requestAuto: true,
-        requestService: {
-          url: '/stream/list',
-          method: 'POST',
-          data: {
-            pageNum: 1,
-            pageSize: 1000,
-            inlongGroupId: values.inlongGroupId,
-          },
-        },
-        requestParams: {
-          formatResult: result =>
-            result?.list.map(item => ({
-              label: item.inlongStreamId,
-              value: item.inlongStreamId,
-            })) || [],
-        },
-      },
-    }),
+    type: 'inputsearch',
+    label: i18n.t('pages.ModuleAudit.config.Ip'),
+    name: 'ip',
   },
   {
     type: 'datepicker',
@@ -107,7 +67,7 @@ export const getFormContent = initialValues => [
     props: {
       allowClear: false,
       showTime: true,
-      format: 'YYYY-MM-DD HH:mm:ss',
+      format: 'YYYY-MM-DD HH:MM:ss',
     },
   },
   {
@@ -118,12 +78,12 @@ export const getFormContent = initialValues => [
     props: {
       allowClear: false,
       showTime: true,
-      format: 'YYYY-MM-DD HH:mm:ss',
+      format: 'YYYY-MM-DD HH:MM:ss',
     },
   },
   {
     type: 'select',
-    label: i18n.t('pages.ModuleAuditDashboard.config.BenchmarkIndicator'),
+    label: i18n.t('pages.ModuleAudit.config.BenchmarkIndicator'),
     name: 'benchmark',
     props: {
       allowClear: true,
@@ -146,7 +106,7 @@ export const getFormContent = initialValues => [
   },
   {
     type: 'select',
-    label: i18n.t('pages.ModuleAuditDashboard.config.ComparativeIndicators'),
+    label: i18n.t('pages.ModuleAudit.config.ComparativeIndicators'),
     name: 'compared',
     props: {
       allowClear: true,
@@ -177,8 +137,16 @@ export const getTableColumns = source => {
   }));
   return [
     {
-      title: i18n.t('pages.ModuleAuditDashboard.config.Ip'),
-      dataIndex: 'ip',
+      title: i18n.t('pages.GroupDetail.Audit.Time'),
+      dataIndex: 'logTs',
+    },
+    {
+      title: i18n.t('pages.ModuleAudit.config.InlongGroupId'),
+      dataIndex: 'inlongGroupId',
+    },
+    {
+      title: i18n.t('pages.ModuleAudit.config.InlongStreamId'),
+      dataIndex: 'inlongStreamId',
     },
   ].concat(data);
 };
